refactor: extract crawling code
